pyproject.toml
[tool.poetry]
name = "text-to-image"
version = "0.1.3"
description = "Generate images from text"
authors = ["Kodai YAMASHITA <nglcobdai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "^2.4.0"
pytz = "^2024.1"
nglcobdai-utils = {git = "https://github.com/nglcobdai/nglcobdai-utils.git", rev = "5cec1b107453f33b5ad31ab33108e2300f7d0f25"}
diffusers = {extras = ["torch"], version = "^0.31.0"}
transformers = "^4.44.2"
bitsandbytes = "^0.44.1"
protobuf = "^5.28.3"
sentencepiece = "^0.2.0"
optimum-quanto = "^0.2.6"
pynvml = "^11.5.3"


[tool.poetry.group.dev.dependencies]
pep8 = "^1.7.1"
black = "^23.9.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
ipython = "^8.15.0"
ipykernel = "^6.25.2"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
